move configs to componenets
